Tweak tests and fix an include dependency
