+++ inc/Tecs_permissions.hh
@@ -1,6 +1,5 @@
 #pragma once
 
-#include <string>
 #include <tuple>
 #include <type_traits>
