+++ tests/test_components.hh
@@ -1,8 +1,8 @@
 #pragma once
 
 #include "test_ecs.hh"
+#include "transform_component.hh"
 
-#include <Tecs.hh>
 #include <cstdint>
 #include <memory>
 #include <string>
@@ -10,29 +10,6 @@
 #include <vector>
 
 namespace testing {
-    struct Transform {
-        double pos[3] = {0};
-        Tecs::Entity parent;
-
-        Transform() {}
-        Transform(double x, double y, double z) : parent() {
-            pos[0] = x;
-            pos[1] = y;
-            pos[2] = z;
-        }
-        Transform(double x, double y, double z, Tecs::Entity parent) : parent(parent) {
-            pos[0] = x;
-            pos[1] = y;
-            pos[2] = z;
-        }
-
-        bool HasParent(Tecs::Lock<ECS, Tecs::Read<Transform>> lock);
-
-        inline bool operator==(const Transform &other) const {
-            return pos[0] == other.pos[0] && pos[1] == other.pos[1] && pos[2] == other.pos[2] && parent == other.parent;
-        }
-    };
-
     struct Script {
         std::vector<uint8_t> data;
         std::shared_ptr<std::string> filename;
