+++ tests/transform_component.cpp
@@ -1,11 +1,6 @@
+#include "transform_component.hh"
+
 #include "test_components.hh"
-
-#include "test_ecs.hh"
-
-#include <Tecs.hh>
-#include <cstdint>
-#include <string>
-#include <vector>
 
 namespace testing {
     bool Transform::HasParent(Tecs::Lock<ECS, Tecs::Read<Transform>> lock) {
